etudes v0.0.1 AskArea

Prompts user for a shape and returns the area.

Functions

area()

Specs

area :: number

Prompts user for a shape and dimensions, returns the area of the shape.

calculate(shape, d1, d2)

Specs

calculate(:atom, Integer, Integer) :: Integer

Given a shape and two dimensions, calculates the area of the given shape.

char_to_shape(arg1)

Specs

char_to_shape(char) :: atom

Given a character returns atom of relevant shape, or unknown.

get_dimensions(atom)

Specs

get_dimensions(atom) :: {number, number}

Given two dimensions, prompts for each dimension. Returns a tuple of the dimensions.

get_number(prompt)

Specs

get_number(String.t) :: number

Offers a prompt and returns a number.

get_shape()

Specs

get_shape :: atom

Offers a prompt and calls the char_to_shape() function to return an the relevant atom.