etudes v0.0.1 College

Étude 7-1: Creating a HashDict from a File

Your local college has given you a text file that contains data about which courses are taught in which rooms. The first column is the course ID number. The second column is the course name, and the third column is the room number.

Your job in this étude is to read the file and create a HashDict whose key is the room number and whose value is a list of all the courses taught in that room.

Functions

make_room_list(fin)

Specs

make_room_list(String.t) :: HashDict.t

Open given courses CSV file, with course ID, course, and room columns. Build a HashDict containing the rooms and all courses associated with each room.