etudes v0.0.1 Geography

Étude 7-2: Creating Structures from a File

In this étude, you will read a CSV file with information about countries and cities. Lines with two entries give a country name and its primary language; lines with four entries give a city name, its population, its latitude, and its longitude.

Functions

make_geo_list(fin)

Specs

make_geo_list(String.t) :: Country.t

Given the name of a CSV file, returns a list of Country structures and their cities.

total_population(geo_list, language)

Specs

total_population([Country], String.t) :: integer

The first argument will be a list as constructed by make_geo_list/1, and the second argument is a string giving the name of a language. The function returns the total population of all the cities in countries whose primary language is the one you specified.