etudes v0.0.1 Geom

Functions for Geometric calculations.

from Études for Elixir, O’Reilly Media, Inc., 2013. Copyright 2013 by J. David Eisenberg.

http://chimera.labs.oreilly.com/books/1234000001642

Functions

area(shape, a, b)

Specs

area(atom, number, number) :: number

Calculates the area of a shape, given the shape and two of the dimensions as a tuple. Returns the productof its arguments for a rectangle, one half the product of the arguments for a triangle, and :math.pi times the product of the arguments for an ellipse. Any invalid data returns zero.