etudes v0.0.1 Powers

Function for raising a number to an integer power.

from Études for Elixir, O’Reilly Media, Inc., 2013. Copyright 2013 by J. David Eisenberg.

http://chimera.labs.oreilly.com/books/1234000001642

Functions

nth_root(x, n)

Specs

nth_root(number, number) :: number

Given x and n, return the nth root of x. This uses the Newton-Raphson method for calculating roots. (See http://en.wikipedia.org/wiki/Newton%27s_method for details).

raise(x, n)

Specs

raise(number, number) :: number

Raise a number x to an integer power n. Any number to the power 0 equals 1. Any number to the power 1 is that number itself. When n is positive, x^n is equal to x times x^(n - 1) When n is negative, x^n is equal to 1.0 / x^n