etudes v0.0.1 Teeth

Functions to ascertain the health of teeth. Dentists check the health of your gums by checking the depth of the “pockets” at six different locations around each of your 32 teeth. The depth is measured in millimeters. If any of the depths is greater than or equal to four millimeters, that tooth needs attention.

Functions

alert(depths)

Specs

alert([list]) :: list

Given a list of lists of pocket depths, returns a list of teeth that require attention.

pocket_depths()